Move e2iphi.h under Platforms/CPU
