+++ src/Platforms/CPU/tests/CMakeLists.txt
@@ -16,7 +16,7 @@ SET(SRC_DIR cpu)
 SET(UTEST_EXE test_${SRC_DIR})
 SET(UTEST_NAME deterministic-unit_test_${SRC_DIR})
 
-ADD_EXECUTABLE(${UTEST_EXE} test_aligned_allocator.cpp)
-TARGET_LINK_LIBRARIES(${UTEST_EXE} platform_cpu catch_main)
+ADD_EXECUTABLE(${UTEST_EXE} test_aligned_allocator.cpp test_e2iphi.cpp)
+TARGET_LINK_LIBRARIES(${UTEST_EXE} platform_device catch_main)
 
 ADD_UNIT_TEST(${UTEST_NAME} "${QMCPACK_UNIT_TEST_DIR}/${UTEST_EXE}")
