+++ src/QMCHamiltonians/MomentumEstimator.cpp
@@ -15,7 +15,7 @@
 
 #include <QMCHamiltonians/MomentumEstimator.h>
 #include <QMCWaveFunctions/TrialWaveFunction.h>
-#include <Numerics/e2iphi.h>
+#include <CPU/e2iphi.h>
 #include <CPU/BLAS.hpp>
 #include <OhmmsData/AttributeSet.h>
 #include <Utilities/SimpleParser.h>
