+++ src/QMCWaveFunctions/EinsplineSetBuilder_createSPOs.cpp
@@ -17,7 +17,7 @@
 //////////////////////////////////////////////////////////////////////////////////////
 
 
-#include "Numerics/e2iphi.h"
+#include "CPU/e2iphi.h"
 #include "CPU/SIMD/vmath.hpp"
 #include "QMCWaveFunctions/EinsplineSetBuilder.h"
 #include "OhmmsData/AttributeSet.h"
